Chore: 채용지원 기능 코드 수정
- 채용지원관련 클래스를 채용공고 클래스에 병합
