+++ kihong1602/recrute/src/main/java/com/blanc/recrute/recruitment/dao/RecruitDAO.java
@@ -1,6 +1,7 @@
 package com.blanc.recrute.recruitment.dao;
 
 import com.blanc.recrute.mybatis.MybatisConnectionFactory;
+import com.blanc.recrute.recruitment.dto.ApplyDTO;
 import com.blanc.recrute.recruitment.dto.DetailDTO;
 import com.blanc.recrute.recruitment.dto.RecruitDTO;
 import org.apache.ibatis.session.SqlSession;
@@ -16,6 +17,24 @@ public class RecruitDAO {
 
         sqlSession.close();
         return detailDTO;
+    }
+
+    public int applyRecruit(ApplyDTO applyDTO) {
+        connSql();
+
+        int result = sqlSession.insert("applyRecruit", applyDTO);
+
+        sqlSession.close();
+        return result;
+    }
+
+    public Integer searchMemberId(String memberRealId) {
+        connSql();
+
+        Integer id = sqlSession.selectOne("searchMemberRealId", memberRealId);
+
+        sqlSession.close();
+        return id;
     }
 
     private void connSql() {
