--- kihong1602/recrute/src/main/resources/mapper/applicant.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="applicant">
-    <insert id="applyRecruit" parameterType="ApplyDTO">
-        insert into applicant (apt_id,member_id,recruitment_id)
-        values (#{aptId},#{memberId},#{recruitId})
-    </insert>
-    <select id="searchMemberRealId" parameterType="String" resultType="Integer">
-        select id from member where member_id = #{memberId}
-    </select>
-</mapper>
